server/requirements.txt
Flask==2.3.2
Pillow==10.0.0
opencv-python-headless==4.8.0.76